\name{GRCh38}
\alias{GRCh38}
\title{Length of human chromosomes based on GRCh38 assembly}
\source{See \url{https://www.ncbi.nlm.nih.gov/assembly/883148/}.}
\description{Length of human chromosomes based on GRCh38 assembly.}
\value{A Data Frame containing length of chromosomes of GRCh38 assembly.}
